% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.r
\name{evaluation}
\alias{evaluation}
\title{generate evaluation metrics based on predicted values}
\usage{
evaluation(y, ypred, metrics = c("MSE", "MAE", "RMSE", "accuracy"))
}
\arguments{
\item{y}{actual response variable.}

\item{ypred}{predicted response variable.}

\item{metrics}{can be MSE, MAE, RMSE and accuracy. accuracy is only applicable to binary responses.}
}
\value{
evaluation metrics.
}
\description{
This functions generate evaluation metrics depending on observed responses and predicted response values.
}
\seealso{
\code{\link{extendedglmnet}}, \code{\link{predict.extendglmnet}} and \code{\link{print.extendglmnet}} methods.
}
\author{
Weiwei Tao \email{tww101928@gmail.com}
}
