% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.extendglmnet}
\alias{print.extendglmnet}
\alias{extendedglmnet.print}
\title{print an extendedglmnet fitting results}
\usage{
\method{extendedglmnet}{print}(x)
}
\arguments{
\item{fit}{fitted extendedglmnet object}
}
\description{
print an extendedglmnet fitting results
}
\details{
The call that the fitting results from extendedglmnet is printed.
The coefficients of the fit followed by degrees of freedom (nobs - 1) and
the residual degree of freedom (nobs - number of nonzero coefficients).
}
\examples{

size = 30
x = matrix(runif(35*size),ncol=35)
y = 10*x[,1] + 3*x[,2] + 20*(x[,3]-0.5)**2 + 10*x[,4] + 5*x[,35] + runif(1,0,1)
fit.regression <- extendedglmnet(x, y, family = "gaussian", type = "regression")
print.extendglmnet(fit.regression)

}
\seealso{
\code{\link{extendedglmnet}}, \code{\link{predict.extendglmnet}} and \code{\link{evaluation}} methods.
}
\author{
Weiwei Tao \email{tww101928@gmail.com}
}
