% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedglmnet.R
\name{extendedglmnet}
\alias{extendedglmnet}
\title{Extended Generalized Linear Models}
\usage{
extendedglmnet(
  trainX,
  trainY,
  response = c("linear", "logistic"),
  model = c("lasso", "ridge", "random lasso"),
  B = 500,
  q1,
  q2
)
}
\arguments{
\item{trainX}{Input matrix of dimension (n observations) x (n variables) where each row is an observation.  Number of parameters must be 2 or more}

\item{trainY}{Response variable containing a vector of n observations.  Elements should be continuous for response="linear" and should contain 0 or 1 for response="logistic"}

\item{response}{String representing type of response variable.  Must be either "linear" or "logistic".  Defaults to "linear"}

\item{model}{String representing model type.  Must be "lasso", "ridge", or "random lasso".  Defaults to "lasso"}

\item{B}{Integer representing number of samples drawn in random lasso.  Defaults to 500}

\item{q1}{Integer representing number of parameters sampled when measuring importance in random lasso.}

\item{q2}{Integer representing number of parameters sampled when calculating final parameters in random lasso.}
}
\description{
This function allows you to train a linear model using several
different algorithms, including lasso, ridge, and random lasso.  It
also allows you to use two different types of response variables,
logistic and linear.
}
